Move unused actions facade APIs out of the way and delete unused juju client code
